Integrate Chat component to Frontend and bug fixes
