+++ frontend/src/App.js
@@ -7,8 +7,6 @@ import SearchPage from './components/SearchPage'
 import InterviewPage from "./components/InterviewPage"
 import SettingsPage from './components/SettingsPage'
 import HistoryPage from "./components/HistoryPage"
-import Chat from "./components/Chat"
-// Ignore Chat route, will remove when integrating to interview.js
 
 function App() {
     return (
@@ -24,8 +22,6 @@ function App() {
                         <Route path="/search" element={<SearchPage />} />
                         <Route path="/interview" element={<InterviewPage />} />
                         <Route path="/history" element={<HistoryPage />} />
-                        <Route path="/room/username1" element={<Chat user={"username1"}/>} />
-                        <Route path="/room/username2" element={<Chat user={"username2"}/>} />
                     </Routes>
                 </Router>
             </Container>
