+++ frontend/src/components/InterviewPage.js
@@ -4,7 +4,10 @@ import { Button, Modal } from 'react-bootstrap'
 import ReactMarkdown from "react-markdown"
 import MatchingService from "../services/matchingService.js"
 import HistoryService from "../services/historyService.js"
-
+import io from 'socket.io-client';
+import Chat from "./Chat.js"
+import { CHAT_SERVICE_URL } from "../configs"
+const socket = io.connect(CHAT_SERVICE_URL); // Connect to chat server
 
 const InterviewPage = () => {
     const [questionTitle, setQuestionTitle] = useState("")
@@ -18,6 +21,7 @@ const InterviewPage = () => {
     const username = location.state.username
 
     useEffect(() => {
+        socket.emit('join_room', { username, room: interviewId}); // Connect user to chat room
         MatchingService
             .getInterview(interviewId)
             .then(interviewDetails => {
@@ -80,22 +84,23 @@ const InterviewPage = () => {
 
     return (
         <>
-            <div className="d-flex justify-content-center mt-5">{questionTitle}</div>
+            <div className="d-flex justify-content-center mt-5"><h1>{questionTitle}</h1></div>
             <div className="container mt-5">
                 <ReactMarkdown>{questionDescription}</ReactMarkdown>
             </div>
+            <Chat socket={socket} username={username} room={interviewId}/>
             <div className="d-flex justify-content-center mt-5">
                 <Button variant="danger" onClick={handleShowEndInterview}>End interview</Button>
                 <Modal className="deleteModal" show={showEndInterview} onHide={handleCloseEndInterview} keyboard={false} animation={false}>
-                <Modal.Header closeButton>
-                    <Modal.Title>Delete</Modal.Title>
-                </Modal.Header>
-                <Modal.Body>Are you sure you want to end the interview?</Modal.Body>
-                <Modal.Footer>
-                    <Button variant="secondary" onClick={handleCloseEndInterview}>Resume interview</Button>
-                    <Button variant="danger" onClick={handleEndInterview}>End interview</Button>
-                </Modal.Footer>
-            </Modal>
+                    <Modal.Header closeButton>
+                        <Modal.Title>Delete</Modal.Title>
+                    </Modal.Header>
+                    <Modal.Body>Are you sure you want to end the interview?</Modal.Body>
+                    <Modal.Footer>
+                        <Button variant="secondary" onClick={handleCloseEndInterview}>Resume interview</Button>
+                        <Button variant="danger" onClick={handleEndInterview}>End interview</Button>
+                    </Modal.Footer>
+                </Modal>
             </div>
         </>
     )
