+++ frontend/src/configs.js
@@ -3,3 +3,5 @@ export const USER_SERVICE_URL = process.env.USER_SERVICE_URL || 'http://localhos
 export const MATCHING_SERVICE_URL = process.env.MATCHING_SERVICE_URL || 'http://localhost:8001'
 
 export const HISTORY_SERVICE_URL = process.env.HISTORY_SERVICE_URL || 'http://localhost:8004'
+
+export const CHAT_SERVICE_URL = process.env.CHAT_SERVICE_URL || 'http://localhost:8005'
